chore: refactor file paths in tracking_object_merger module
