+++ perception/tracking_object_merger/include/autoware/tracking_object_merger/association/solver/mu_ssp.hpp
@@ -12,24 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TRACKING_OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
-#define TRACKING_OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
+#ifndef AUTOWARE__TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__MU_SSP_HPP_
+#define AUTOWARE__TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__MU_SSP_HPP_
+
+#include "autoware/tracking_object_merger/association/solver/gnn_solver_interface.hpp"
 
 #include <unordered_map>
 #include <vector>
 
 namespace gnn_solver
 {
-class GnnSolverInterface
+class MuSSP : public GnnSolverInterface
 {
 public:
-  GnnSolverInterface() = default;
-  virtual ~GnnSolverInterface() = default;
+  MuSSP() = default;
+  ~MuSSP() = default;
 
-  virtual void maximizeLinearAssignment(
+  void maximizeLinearAssignment(
     const std::vector<std::vector<double>> & cost, std::unordered_map<int, int> * direct_assignment,
-    std::unordered_map<int, int> * reverse_assignment) = 0;
+    std::unordered_map<int, int> * reverse_assignment) override;
 };
 }  // namespace gnn_solver
 
-#endif  // TRACKING_OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
+#endif  // AUTOWARE__TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__MU_SSP_HPP_
