+++ perception/tracking_object_merger/include/autoware/tracking_object_merger/decorative_tracker_merger.hpp
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TRACKING_OBJECT_MERGER__DECORATIVE_TRACKER_MERGER_HPP_
-#define TRACKING_OBJECT_MERGER__DECORATIVE_TRACKER_MERGER_HPP_
+#ifndef AUTOWARE__TRACKING_OBJECT_MERGER__DECORATIVE_TRACKER_MERGER_HPP_
+#define AUTOWARE__TRACKING_OBJECT_MERGER__DECORATIVE_TRACKER_MERGER_HPP_
 
-#include "tracking_object_merger/data_association/data_association.hpp"
-#include "tracking_object_merger/utils/tracker_state.hpp"
-#include "tracking_object_merger/utils/utils.hpp"
+#include "autoware/tracking_object_merger/association/data_association.hpp"
+#include "autoware/tracking_object_merger/utils/tracker_state.hpp"
+#include "autoware/tracking_object_merger/utils/utils.hpp"
+#include "autoware/universe_utils/ros/debug_publisher.hpp"
+#include "autoware/universe_utils/ros/published_time_publisher.hpp"
+#include "autoware/universe_utils/system/stop_watch.hpp"
 
-#include <autoware/universe_utils/ros/debug_publisher.hpp>
-#include <autoware/universe_utils/ros/published_time_publisher.hpp>
-#include <autoware/universe_utils/system/stop_watch.hpp>
 #include <rclcpp/rclcpp.hpp>
 
 #include <autoware_perception_msgs/msg/tracked_objects.hpp>
@@ -133,4 +133,4 @@ private:
 
 }  // namespace tracking_object_merger
 
-#endif  // TRACKING_OBJECT_MERGER__DECORATIVE_TRACKER_MERGER_HPP_
+#endif  // AUTOWARE__TRACKING_OBJECT_MERGER__DECORATIVE_TRACKER_MERGER_HPP_
